RF: move run_test logic which should be common into _abstract_runner
Also declare @abc.abstractmethod _run_test for specific
implementations, implemented by pytest plugin
